Login Functionality
Created Customer Login and Staff Login and also made it so that when a user tries to login and their username is in the database then it says that their password is incorrect. If the user is not in the database then it asks them if they would like to register.